DoctorCancunAgentNanoBot/pyproject.toml
[tool.poetry]
name = "doctor-cancun-agent-nanobot"
version = "0.1.0"
description = "AI chatbot agent for iDoctor Cancún - phone repair shop"
authors = ["Julian Juarez <sistemascancunjefe@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
pyyaml = "^6.0"
httpx = "^0.28.0"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.0"
aiosqlite = "^0.20.0"
python-telegram-bot = "^21.0"
apscheduler = "^3.10.0"
anthropic = "^0.42.0"
openai = "^1.60.0"
groq = "^0.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24.0"
black = "^24.0"
ruff = "^0.8.0"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
